requirements.txt
addict==2.4.0
matplotlib==3.7.2
numba==0.57.1
numpy==1.24.4
open3d==0.17.0
pandas==2.0.3
pytorch3d==0.7.4
PyYAML==6.0
scikit_learn==1.3.0
spconv_cu117==2.3.6
torch==2.0.1
tqdm==4.65.0
xgboost==1.7.6
